Refactor: Merge small modules to reduce fragmentation
- Merged `src/semantic/types.rs` into `src/semantic/model.rs`
- Merged `src/morphology/matcher.rs` into `src/morphology/mod.rs`
- Merged `src/parser/recursion.rs` into `src/parser/mod.rs`
- Deleted source files
- Updated usages
- Documented in .jules/razor.md
